fix(xhci): use u32 for interrupt transfer data_len to match bulk API
Consistent with bulk_transfer_in/out which also take u32. The TRB
field is 17-bit so the runtime check in enqueue_normal catches
oversized values; no need for type-level narrowing.

+++ crates/harmony-unikernel/src/drivers/dwc_usb/mod.rs
@@ -994,7 +994,7 @@ impl XhciDriver {
         slot_id: u8,
         endpoint_id: u8,
         data_buf_phys: u64,
-        data_len: u16,
+        data_len: u32,
     ) -> Result<Vec<XhciAction>, XhciError> {
         if endpoint_id % 2 == 0 || endpoint_id < 2 {
             return Err(XhciError::InvalidState);
@@ -1003,7 +1003,7 @@ impl XhciDriver {
             slot_id,
             endpoint_id,
             data_buf_phys,
-            data_len as u32,
+            data_len,
             trb::IOC | trb::ISP,
         )
     }
@@ -1017,7 +1017,7 @@ impl XhciDriver {
         slot_id: u8,
         endpoint_id: u8,
         data_buf_phys: u64,
-        data_len: u16,
+        data_len: u32,
     ) -> Result<Vec<XhciAction>, XhciError> {
         if endpoint_id % 2 != 0 || endpoint_id < 2 {
             return Err(XhciError::InvalidState);
@@ -1026,7 +1026,7 @@ impl XhciDriver {
             slot_id,
             endpoint_id,
             data_buf_phys,
-            data_len as u32,
+            data_len,
             trb::IOC,
         )
     }
